requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.7
botocore==1.35.7
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.8.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==6.2.11
environs==9.5.0
exa-py==1.0.18
exceptiongroup==1.2.2
fastapi==0.112.2
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
groq==0.10.0
grpcio==1.63.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-anthropic==0.1.23
langchain-cohere==0.2.3
langchain-community==0.2.12
langchain-core==0.2.35
langchain-experimental==0.0.64
langchain-google-genai==1.0.10
langchain-groq==0.1.9
langchain-milvus==0.1.4
langchain-nvidia-ai-endpoints==0.2.1
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.106
markdown-it-py==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
milvus-lite==2.4.9
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.42.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
primp==0.6.1
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.7
pydantic_core==2.20.1
Pygments==2.18.0
pymilvus==2.4.5
pyparsing==3.1.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rsa==4.9
s3transfer==0.10.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.38.2
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
yarl==1.9.4
